requirements.txt
streamlit>=1.40.0
deepgram-sdk>=3.5.0
elevenlabs>=1.0.0
anthropic>=0.40.0
python-dotenv>=1.0.0
fpdf2>=2.8.0
